# Remove bool_to_int