This moves a bit of complexity from struct_deserializer into
value_serializer(:bool). Things are a little more coherent overall, and it also
helps simplify struct_deserializer, which is getting complex.