@pinterest-ebert 2016-12-27
Ebert has finished reviewing this Pull Request and has found:

* 1 fixed issue! 🎉

You can see more details about this review at https://ebertapp.io/github/pinterest/elixir-thrift/pulls/89.

<!-- Ebert auto generated comment -->