@pguillory 2016-12-27
Seems to be a smidge faster. I ran the benchmark a couple of times before and after the change.

```
before:
elixir serialization (left as IOList)             5000   903.73 µs/op
elixir serialization (iolist_size)                5000   990.38 µs/op
elixir serialization (converted to binary)        5000   1317.05 µs/op

elixir serialization (left as IOList)             5000   901.92 µs/op
elixir serialization (iolist_size)                5000   978.69 µs/op
elixir serialization (converted to binary)        5000   1342.56 µs/op

after:
elixir serialization (left as IOList)             5000   857.57 µs/op
elixir serialization (iolist_size)                5000   961.09 µs/op
elixir serialization (converted to binary)        5000   1307.12 µs/op

elixir serialization (left as IOList)             5000   858.72 µs/op
elixir serialization (iolist_size)                5000   960.23 µs/op
elixir serialization (converted to binary)        5000   1282.42 µs/op
```